+++ Documentation/SwiftUI MVVM Architecture/Example Project/SwiftUIMVVMExample/Models/Post.swift
@@ -1,0 +1,49 @@
+//
+//  Post.swift
+//  SwiftUIMVVMExample
+//
+//  Created by Vakhtang Kontridze on 19.06.22.
+//
+
+import Foundation
+
+// MARK: - Post
+struct Post: Identifiable, Equatable, Hashable {
+    // MARK: Properties
+    let id: Int
+    let title: String
+    let body: String
+    
+    // MARK: Mock
+    static var mock: Self {
+        .init(
+            id: 1,
+            title: "Lorem Ipsum",
+            body: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam imperdiet augue eget odio posuere pharetra. Donec et vestibulum turpis. Integer consequat sapien ut ligula facilisis, nec aliquet erat aliquet. Praesent sed scelerisque enim. Cras quam nibh, tincidunt non sollicitudin non, egestas in mi. Morbi tempor sit amet ligula nec lacinia. Curabitur malesuada placerat lectus, sed rutrum sapien semper id. Aliquam ac sagittis urna. Nam pharetra nulla et nunc mollis, vel molestie justo consequat. Praesent ac ullamcorper quam. Duis sagittis mauris quam, vel maximus ligula ultrices eu. Proin massa ligula, ornare sit amet convallis sed, tincidunt nec neque."
+        )
+    }
+
+    // MARK: Hashable
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(id)
+    }
+}
+
+// MARK: - Mapping
+extension Post {
+    init?(entity: PostsEntity.Post) {
+        guard
+            let id: Int = entity.id,
+            let title: String = entity.title,
+            let body: String = entity.body
+        else {
+            return nil
+        }
+
+        self.init(
+            id: id,
+            title: title.capitalized,
+            body: body
+        )
+    }
+}