+++ Documentation/SwiftUI MVVM Architecture/Example Project/SwiftUIMVVMExample/Scenes/Posts/PostsParameters.swift
@@ -1,0 +1,11 @@
+//
+//  PostsParameters.swift
+//  SwiftUIMVVMExample
+//
+//  Created by Vakhtang Kontridze on 20.06.22.
+//
+
+import Foundation
+
+// MARK: - Post Parameters
+struct PostParameters: Hashable {}