+++ Sources/VCore/Extensions/SwiftUI/View/View.OnFirstAppear.swift
@@ -11,21 +11,11 @@ import SwiftUI
 extension View {
     /// Adds an action to perform before this `View` appears for the first time.
     ///
-    ///     struct ContentView: View {
-    ///         @StateObject private var presenter: Presenter = .init()
+    ///     @State private var didAppearForTheFirstTime: Bool = false
     ///
-    ///         @State private var didAppearForTheFirstTime: Bool = false
-    ///
-    ///         var body: some View {
-    ///             content
-    ///                 .onFirstAppear($didAppearForTheFirstTime, perform: { presenter.didLoad() })
-    ///         }
-    ///     }
-    ///
-    ///     @MainActor final class Presenter: ObservableObject {
-    ///         func didLoad() {
-    ///             ...
-    ///         }
+    ///     var body: some View {
+    ///         content
+    ///             .onFirstAppear($didAppearForTheFirstTime, perform: { ... })
     ///     }
     ///
     public func onFirstAppear(