+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/CoordinatingNavigationStack.swift
@@ -11,55 +11,21 @@ import SwiftUI
 /// `NavigationStack` that manages `NavigationPath` for representing content in the stack.
 ///
 /// `View` embeds `NavigationStackCoordinator` to the environment for it to be used by contents in the stack.
-/// Can be used to pass `NavigationStackCoordinator` to `Presenter` or `ViewModel` for pushing or popping.
+/// Can be used to push or pop `View`s programmatically.
 ///
-///     struct HomeView: View {
-///         @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator?
-///         @StateObject private var presenter: HomePresenter
-///         @State private var didAppearForTheFirstTime: Bool = false
+///     @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator!
 ///
-///         var body: some View {
-///             CoordinatingNavigationStack(root: {
-///                 Button(
-///                     "Lorem ipsum",
-///                     action: { presenter.didTapButton() }
-///                 )
-///                 .modifier(HomeRouter())
-///                 .onFirstAppear($didAppearForTheFirstTime, perform: {
-///                     presenter.navigationStackCoordinator = navigationStackCoordinator
-///                 })
-///             })
-///         }
-///     }
-///
-///     struct HomeRouter: ViewModifier {
-///         func body(content: Content) -> some View {
-///             content
-///                 .navigationDestination(
-///                     for: DestinationParameters.self,
-///                     destination: { DestinationView(parameters: $0) }
-///                 )
-///         }
-///     }
-///
-///     final class HomePresenter: ObservableObject {
-///         @Published var navigationStackCoordinator: NavigationStackCoordinator?
-///
-///         func didTapButton() {
-///             navigationStackCoordinator?.path.append(DestinationParameters())
-///         }
-///     }
-///
-///     struct DestinationParameters: Hashable { ... }
-///
-///     struct DestinationView: View {
-///         init(parameters: DestinationParameters) {
-///             ...
-///         }
-///
-///         var body: some View {
-///             ...
-///         }
+///     var body: some View {
+///         CoordinatingNavigationStack(root: {
+///             Button(
+///                 "Lorem ipsum",
+///                 action: { navigationStackCoordinator.path.append(DestinationParameters()) }
+///             )
+///             .navigationDestination(
+///                 for: DestinationParameters.self,
+///                 destination: { DestinationView(parameters: $0) }
+///             )
+///         })
 ///     }
 ///
 @available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)