+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/NavigationStackCoordinator.swift
@@ -8,57 +8,23 @@
 import SwiftUI
 
 // MARK: - Navigation Stack Coordinator
-/// `ObservableObject` that stores `NavigationPath` for coordinating with `CoordinatingNavigationStack`.
+/// Object that stores `NavigationPath` for coordinating with `CoordinatingNavigationStack`.
 ///
-/// Can be used to push or pop `View`s from `Presenter` or `ViewModel`.
+/// Can be used to push or pop `View`s programmatically.
 ///
-///     struct HomeView: View {
-///         @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator?
-///         @StateObject private var presenter: HomePresenter
-///         @State private var didAppearForTheFirstTime: Bool = false
+///     @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator!
 ///
-///         var body: some View {
-///             CoordinatingNavigationStack(root: {
-///                 Button(
-///                     "Lorem ipsum",
-///                     action: { presenter.didTapButton() }
-///                 )
-///                 .modifier(HomeRouter())
-///                 .onFirstAppear($didAppearForTheFirstTime, perform: {
-///                     presenter.navigationStackCoordinator = navigationStackCoordinator
-///                 })
-///             })
-///         }
-///     }
-///
-///     struct HomeRouter: ViewModifier {
-///         func body(content: Content) -> some View {
-///             content
-///                 .navigationDestination(
-///                     for: DestinationParameters.self,
-///                     destination: { DestinationView(parameters: $0) }
-///                 )
-///         }
-///     }
-///
-///     final class HomePresenter: ObservableObject {
-///         @Published var navigationStackCoordinator: NavigationStackCoordinator?
-///
-///         func didTapButton() {
-///             navigationStackCoordinator?.path.append(DestinationParameters())
-///         }
-///     }
-///
-///     struct DestinationParameters: Hashable { ... }
-///
-///     struct DestinationView: View {
-///         init(parameters: DestinationParameters) {
-///             ...
-///         }
-///
-///         var body: some View {
-///             ...
-///         }
+///     var body: some View {
+///         CoordinatingNavigationStack(root: {
+///             Button(
+///                 "Lorem ipsum",
+///                 action: { navigationStackCoordinator.path.append(DestinationParameters()) }
+///             )
+///             .navigationDestination(
+///                 for: DestinationParameters.self,
+///                 destination: { DestinationView(parameters: $0) }
+///             )
+///         })
 ///     }
 ///
 @available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)