+++ Sources/VCore/Helpers/Architectural Pattern Helpers/UIKit/Alert/UIAlertViewable.swift
@@ -38,7 +38,6 @@ import UIKit
 ///         }
 ///     }
 ///
-///     
 public protocol UIAlertViewable {
     /// Presents `UIAlert` with parameters
     func presentAlert(parameters: UIAlertParameters)